+++ server/api/initiate-login-as.js
@@ -7,7 +7,12 @@ const USING_SSL = process.env.REACT_APP_SHARETRIBE_USING_SSL === 'true';
 
 // redirect_uri param used when initiating a login as authentication flow and
 // when requesting a token using an authorization code
-const loginAsRedirectUri = `${ROOT_URL.replace(/\/$/, '')}/api/login-as`;
+const loginAsRedirectUri = redirect_path_after_login => {
+  const query = redirect_path_after_login
+    ? `?redirect_path_after_login=${redirect_path_after_login}`
+    : '';
+  return `${ROOT_URL.replace(/\/$/, '')}/api/login-as${query}`;
+};
 
 // Cookies used for authorization code authentication.
 const stateKey = `st-${CLIENT_ID}-oauth2State`;
@@ -52,11 +57,12 @@ module.exports = (req, res) => {
     .digest('base64');
   const codeChallenge = urlifyBase64(hash);
   const authorizeServerUrl = `${CONSOLE_URL}/api/authorize-as`;
+  const { redirect_path_after_login } = req.query || {};
 
   const location = `${authorizeServerUrl}?\
 response_type=code&\
 client_id=${CLIENT_ID}&\
-redirect_uri=${loginAsRedirectUri}&\
+redirect_uri=${loginAsRedirectUri(redirect_path_after_login)}&\
 user_id=${userId}&\
 state=${state}&\
 code_challenge=${codeChallenge}&\